requirements.txt
asgiref==3.5.2
Django==3.2
pytz==2022.1
sqlparse==0.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0